+++ src/main/java/cn/projectan/strix/core/module/oauth/WechatOAuthClient.java
@@ -66,7 +66,7 @@ public class WechatOAuthClient extends StrixOAuthClient {
     }
 
     @Override
-    public int getPlatform() {
+    public short getPlatform() {
         return config.getPlatform();
     }
